POC for new CLI configuration
