+++ packages/cli-lib/lib/constants.js
@@ -39,6 +39,8 @@ const Mode = {
 const DEFAULT_MODE = Mode.publish;
 
 const DEFAULT_HUBSPOT_CONFIG_YAML_FILE_NAME = 'hubspot.config.yml';
+const HUBSPOT_CONFIGURATION_FOLDER = '.hubspot';
+const HUBSPOT_CONFIGURATION_FILE = 'config.yml';
 
 const EMPTY_CONFIG_FILE_CONTENTS = '';
 
@@ -86,6 +88,7 @@ const ENVIRONMENT_VARIABLES = {
   HUBSPOT_CLIENT_SECRET: 'HUBSPOT_CLIENT_SECRET',
   HUBSPOT_PERSONAL_ACCESS_KEY: 'HUBSPOT_PERSONAL_ACCESS_KEY',
   HUBSPOT_PORTAL_ID: 'HUBSPOT_PORTAL_ID',
+  HUBSPOT_ACCOUNT_ID: 'HUBSPOT_ACCOUNT_ID',
   HUBSPOT_REFRESH_TOKEN: 'HUBSPOT_REFRESH_TOKEN',
   HUBSPOT_ENVIRONMENT: 'HUBSPOT_ENVIRONMENT',
 };
@@ -236,6 +239,8 @@ module.exports = {
   API_KEY_AUTH_METHOD,
   AUTH_METHODS,
   DEFAULT_HUBSPOT_CONFIG_YAML_FILE_NAME,
+  HUBSPOT_CONFIGURATION_FOLDER,
+  HUBSPOT_CONFIGURATION_FILE,
   DEFAULT_MODE,
   DEFAULT_OAUTH_SCOPES,
   ENVIRONMENT_VARIABLES,
